refactor: Update useFieldSelector structure
