+++ theme/src/recipes/input/createFieldRecipe.test.ts
@@ -9,6 +9,7 @@ import {
 	createFieldGroupRecipe,
 	createFieldRecipe,
 	createFieldSlotRecipe,
+	useFieldSelector,
 } from "./createFieldRecipe";
 
 function createInstance() {
@@ -61,6 +62,7 @@ function createInstance() {
 		s.utility(name, ({ value }) => ({ [name]: value }));
 	}
 	s.variable("color.text-weakest", "#64748B", { default: true });
+
 	useDarkModifier(s);
 	useHoverModifier(s);
 	useFocusWithinModifier(s);
@@ -213,5 +215,48 @@ describe("createFieldGroupRecipe", () => {
 			);
 			expect(rule, `missing seam rule: ${query}`).toBeDefined();
 		}
+	});
+});
+
+describe("useFieldSelector", () => {
+	function findSelector(s: ReturnType<typeof styleframe>, query: string) {
+		return s.root.children.find(
+			(child) =>
+				child.type === "selector" &&
+				(child as { query: string }).query === query,
+		) as
+			| {
+					type: "selector";
+					query: string;
+					declarations: Record<string, unknown>;
+			  }
+			| undefined;
+	}
+
+	it("should register a selector with the given query", () => {
+		const s = createInstance();
+		useFieldSelector(s, ".input-field");
+
+		expect(findSelector(s, ".input-field")).toBeDefined();
+	});
+
+	it("should target whatever field query it is given", () => {
+		const s = createInstance();
+		useFieldSelector(s, ".textarea-field");
+
+		expect(findSelector(s, ".textarea-field")).toBeDefined();
+		expect(findSelector(s, ".input-field")).toBeUndefined();
+	});
+
+	it("should apply a transparent, borderless reset", () => {
+		const s = createInstance();
+		useFieldSelector(s, ".input-field");
+
+		expect(findSelector(s, ".input-field")?.declarations).toMatchObject({
+			width: "100%",
+			background: "transparent",
+			border: "none",
+			padding: "0",
+		});
 	});
 });
