+++ theme/src/recipes/input/createFieldRecipe.ts
@@ -3,319 +3,6 @@ import {
 	type RecipeConfig,
 	createUseRecipe,
 } from "../../utils/createUseRecipe";
-import { useFieldSelector } from "../../utils/useFieldSelector";
-
-/**
- * Shared field-recipe definitions for the Input and Textarea families.
- *
- * Input is the canonical field recipe; Textarea extends the same surface. Both
- * the input and textarea recipe files build their parts from the factories
- * below, so the color/variant/size/state system lives in a single place. These
- * builders are intentionally not re-exported from the package barrel &mdash;
- * they are internal to the field recipes.
- */
-
-/**
- * Common field-wrapper base, minus `display`/`alignItems` (each field sets its
- * own: inline-flex/center for input, flex/flex-start for textarea).
- */
-const fieldRecipeBase = {
-	fontFamily: "inherit",
-	fontSize: "@font-size.sm",
-	fontWeight: "@font-weight.normal",
-	lineHeight: "@line-height.normal",
-	borderWidth: "@border-width.thin",
-	borderStyle: "@border-style.solid",
-	borderColor: "transparent",
-	borderRadius: "@border-radius.md",
-	paddingTop: "@0.5",
-	paddingBottom: "@0.5",
-	paddingLeft: "@0.75",
-	paddingRight: "@0.75",
-	background: "transparent",
-	color: "@color.text",
-	outline: "none",
-	transitionProperty: "color, background-color, border-color",
-	transitionTimingFunction: "@easing.ease-in-out",
-	transitionDuration: "150ms",
-	"&:focus-within": {
-		outlineWidth: "2px",
-		outlineStyle: "solid",
-		outlineColor: "@color.primary",
-		outlineOffset: "2px",
-	},
-};
-
-/** Shared variant axes: color, visual style, size, and the three boolean states. */
-const fieldRecipeVariants = {
-	color: {
-		light: {},
-		dark: {},
-		neutral: {},
-	},
-	variant: {
-		default: {},
-		soft: {},
-		ghost: {},
-	},
-	size: {
-		sm: {
-			fontSize: "@font-size.xs",
-			paddingTop: "@0.375",
-			paddingBottom: "@0.375",
-			paddingLeft: "@0.625",
-			paddingRight: "@0.625",
-			borderRadius: "@border-radius.sm",
-		},
-		md: {
-			fontSize: "@font-size.sm",
-			paddingTop: "@0.5",
-			paddingBottom: "@0.5",
-			paddingLeft: "@0.75",
-			paddingRight: "@0.75",
-			borderRadius: "@border-radius.md",
-		},
-		lg: {
-			fontSize: "@font-size.md",
-			paddingTop: "@0.625",
-			paddingBottom: "@0.625",
-			paddingLeft: "@0.875",
-			paddingRight: "@0.875",
-			borderRadius: "@border-radius.md",
-		},
-	},
-	invalid: {
-		true: {},
-		false: {},
-	},
-	disabled: {
-		true: {},
-		false: {},
-	},
-	readonly: {
-		true: {},
-		false: {},
-	},
-};
-
-/**
- * The 12 compound variants: 9 color×variant combinations plus the invalid,
- * read-only, and disabled state overrides. Order matters &mdash; invalid is
- * placed before read-only and disabled so its error border wins over the
- * color/variant rules but can still be dimmed by the disabled state.
- */
-const fieldRecipeCompoundVariants = [
-	// Light color (fixed across themes)
-	{
-		match: { color: "light", variant: "default" },
-		css: {
-			background: "@color.white",
-			borderColor: "@color.gray-200",
-			color: "@color.text",
-			"&:hover": {
-				borderColor: "@color.gray-300",
-			},
-			"&:dark": {
-				background: "@color.white",
-				borderColor: "@color.gray-200",
-				color: "@color.text-inverted",
-			},
-			"&:dark:hover": {
-				borderColor: "@color.gray-300",
-			},
-		},
-	},
-	{
-		match: { color: "light", variant: "soft" },
-		css: {
-			background: "@color.gray-100",
-			borderColor: "@color.gray-200",
-			color: "@color.text",
-			"&:hover": {
-				borderColor: "@color.gray-300",
-			},
-			"&:dark": {
-				background: "@color.gray-100",
-				borderColor: "@color.gray-200",
-				color: "@color.text-inverted",
-			},
-			"&:dark:hover": {
-				borderColor: "@color.gray-300",
-			},
-		},
-	},
-	{
-		match: { color: "light", variant: "ghost" },
-		css: {
-			background: "transparent",
-			borderColor: "transparent",
-			color: "@color.text",
-			"&:dark": {
-				background: "transparent",
-				borderColor: "transparent",
-				color: "@color.text-inverted",
-			},
-		},
-	},
-
-	// Dark color (fixed across themes)
-	{
-		match: { color: "dark", variant: "default" },
-		css: {
-			background: "@color.gray-900",
-			borderColor: "@color.gray-700",
-			color: "@color.white",
-			"&:hover": {
-				borderColor: "@color.gray-600",
-			},
-			"&:dark": {
-				background: "@color.gray-900",
-				borderColor: "@color.gray-700",
-				color: "@color.white",
-			},
-			"&:dark:hover": {
-				borderColor: "@color.gray-600",
-			},
-		},
-	},
-	{
-		match: { color: "dark", variant: "soft" },
-		css: {
-			background: "@color.gray-800",
-			borderColor: "@color.gray-700",
-			color: "@color.gray-300",
-			"&:hover": {
-				borderColor: "@color.gray-600",
-			},
-			"&:dark": {
-				background: "@color.gray-800",
-				borderColor: "@color.gray-700",
-				color: "@color.gray-300",
-			},
-			"&:dark:hover": {
-				borderColor: "@color.gray-600",
-			},
-		},
-	},
-	{
-		match: { color: "dark", variant: "ghost" },
-		css: {
-			background: "transparent",
-			borderColor: "transparent",
-			color: "@color.gray-300",
-			"&:dark": {
-				background: "transparent",
-				borderColor: "transparent",
-				color: "@color.gray-300",
-			},
-		},
-	},
-
-	// Neutral color (adaptive: light in light mode, dark in dark mode)
-	{
-		match: { color: "neutral", variant: "default" },
-		css: {
-			background: "@color.white",
-			borderColor: "@color.gray-200",
-			color: "@color.text",
-			"&:hover": {
-				borderColor: "@color.gray-300",
-			},
-			"&:dark": {
-				background: "@color.gray-900",
-				borderColor: "@color.gray-700",
-				color: "@color.white",
-			},
-			"&:dark:hover": {
-				borderColor: "@color.gray-600",
-			},
-		},
-	},
-	{
-		match: { color: "neutral", variant: "soft" },
-		css: {
-			background: "@color.gray-100",
-			borderColor: "@color.gray-200",
-			color: "@color.text",
-			"&:hover": {
-				borderColor: "@color.gray-300",
-			},
-			"&:dark": {
-				background: "@color.gray-800",
-				borderColor: "@color.gray-700",
-				color: "@color.gray-300",
-			},
-			"&:dark:hover": {
-				borderColor: "@color.gray-600",
-			},
-		},
-	},
-	{
-		match: { color: "neutral", variant: "ghost" },
-		css: {
-			background: "transparent",
-			borderColor: "transparent",
-			color: "@color.text",
-			"&:dark": {
-				background: "transparent",
-				borderColor: "transparent",
-				color: "@color.white",
-			},
-		},
-	},
-
-	// Invalid state — overrides border and focus ring to error color.
-	{
-		match: { invalid: "true" },
-		css: {
-			borderColor: "@color.error",
-			"&:hover": {
-				borderColor: "@color.error",
-			},
-			"&:focus-within": {
-				outlineColor: "@color.error",
-			},
-			"&:dark": {
-				borderColor: "@color.error",
-			},
-			"&:dark:hover": {
-				borderColor: "@color.error",
-			},
-		},
-	},
-
-	// Read-only state — subtle background shift, default cursor.
-	{
-		match: { readonly: "true" },
-		css: {
-			background: "@color.gray-50",
-			cursor: "default",
-			"&:dark": {
-				background: "@color.gray-800",
-			},
-		},
-	},
-
-	// Disabled state — dim, not-allowed cursor, block interaction.
-	{
-		match: { disabled: "true" },
-		css: {
-			opacity: "0.5",
-			cursor: "not-allowed",
-			pointerEvents: "none",
-		},
-	},
-];
-
-/** Shared default selections (everything except textarea's `resize`). */
-const fieldRecipeDefaultVariants = {
-	color: "neutral",
-	variant: "default",
-	size: "md",
-	invalid: "false",
-	disabled: "false",
-	readonly: "false",
-};
 
 /**
  * Builds a field-wrapper recipe (`.input` / `.textarea`). Merges the shared
@@ -331,7 +18,7 @@ export function createFieldRecipe<
 	const ExtraVariants extends Record<
 		string,
 		Record<string, VariantDeclarationsBlock>
-	> = {},
+	> = Record<string, Record<string, VariantDeclarationsBlock>>,
 >(
 	name: string,
 	extras: {
@@ -342,20 +29,321 @@ export function createFieldRecipe<
 	} = {},
 	setup?: (s: Styleframe) => void,
 ) {
+	/** Shared variant axes: color, visual style, size, and the three boolean states. */
+	const fieldRecipeVariants = {
+		color: {
+			light: {},
+			dark: {},
+			neutral: {},
+		},
+		variant: {
+			default: {},
+			soft: {},
+			ghost: {},
+		},
+		size: {
+			sm: {
+				fontSize: "@font-size.xs",
+				paddingTop: "@0.375",
+				paddingBottom: "@0.375",
+				paddingLeft: "@0.625",
+				paddingRight: "@0.625",
+				borderRadius: "@border-radius.sm",
+			},
+			md: {
+				fontSize: "@font-size.sm",
+				paddingTop: "@0.5",
+				paddingBottom: "@0.5",
+				paddingLeft: "@0.75",
+				paddingRight: "@0.75",
+				borderRadius: "@border-radius.md",
+			},
+			lg: {
+				fontSize: "@font-size.md",
+				paddingTop: "@0.625",
+				paddingBottom: "@0.625",
+				paddingLeft: "@0.875",
+				paddingRight: "@0.875",
+				borderRadius: "@border-radius.md",
+			},
+		},
+		invalid: {
+			true: {},
+			false: {},
+		},
+		disabled: {
+			true: {},
+			false: {},
+		},
+		readonly: {
+			true: {},
+			false: {},
+		},
+	};
+
 	return createUseRecipe(
 		name,
 		{
-			base: { ...fieldRecipeBase, ...extras.base },
+			/**
+			 * Shared field-recipe definitions for the Input and Textarea families.
+			 *
+			 * Input is the canonical field recipe; Textarea extends the same surface. Both
+			 * the input and textarea recipe files build their parts from the factories
+			 * below, so the color/variant/size/state system lives in a single place. These
+			 * builders are intentionally not re-exported from the package barrel &mdash;
+			 * they are internal to the field recipes.
+			 *
+			 * Common field-wrapper base, minus `display`/`alignItems` (each field sets its
+			 * own: inline-flex/center for input, flex/flex-start for textarea).
+			 */
+			base: {
+				fontFamily: "inherit",
+				fontSize: "@font-size.sm",
+				fontWeight: "@font-weight.normal",
+				lineHeight: "@line-height.normal",
+				borderWidth: "@border-width.thin",
+				borderStyle: "@border-style.solid",
+				borderColor: "transparent",
+				borderRadius: "@border-radius.md",
+				paddingTop: "@0.5",
+				paddingBottom: "@0.5",
+				paddingLeft: "@0.75",
+				paddingRight: "@0.75",
+				background: "transparent",
+				color: "@color.text",
+				outline: "none",
+				transitionProperty: "color, background-color, border-color",
+				transitionTimingFunction: "@easing.ease-in-out",
+				transitionDuration: "150ms",
+				"&:focus-within": {
+					outlineWidth: "2px",
+					outlineStyle: "solid",
+					outlineColor: "@color.primary",
+					outlineOffset: "2px",
+				},
+				...extras.base,
+			},
 			variants: {
 				...fieldRecipeVariants,
 				...extras.variants,
 			} as typeof fieldRecipeVariants & ExtraVariants,
+			/**
+			 * The 12 compound variants: 9 color×variant combinations plus the invalid,
+			 * read-only, and disabled state overrides. Order matters &mdash; invalid is
+			 * placed before read-only and disabled so its error border wins over the
+			 * color/variant rules but can still be dimmed by the disabled state.
+			 */
 			compoundVariants: [
-				...fieldRecipeCompoundVariants,
+				// Light color (fixed across themes)
+				{
+					match: { color: "light", variant: "default" },
+					css: {
+						background: "@color.white",
+						borderColor: "@color.gray-200",
+						color: "@color.text",
+						"&:hover": {
+							borderColor: "@color.gray-300",
+						},
+						"&:dark": {
+							background: "@color.white",
+							borderColor: "@color.gray-200",
+							color: "@color.text-inverted",
+						},
+						"&:dark:hover": {
+							borderColor: "@color.gray-300",
+						},
+					},
+				},
+				{
+					match: { color: "light", variant: "soft" },
+					css: {
+						background: "@color.gray-100",
+						borderColor: "@color.gray-200",
+						color: "@color.text",
+						"&:hover": {
+							borderColor: "@color.gray-300",
+						},
+						"&:dark": {
+							background: "@color.gray-100",
+							borderColor: "@color.gray-200",
+							color: "@color.text-inverted",
+						},
+						"&:dark:hover": {
+							borderColor: "@color.gray-300",
+						},
+					},
+				},
+				{
+					match: { color: "light", variant: "ghost" },
+					css: {
+						background: "transparent",
+						borderColor: "transparent",
+						color: "@color.text",
+						"&:dark": {
+							background: "transparent",
+							borderColor: "transparent",
+							color: "@color.text-inverted",
+						},
+					},
+				},
+
+				// Dark color (fixed across themes)
+				{
+					match: { color: "dark", variant: "default" },
+					css: {
+						background: "@color.gray-900",
+						borderColor: "@color.gray-700",
+						color: "@color.white",
+						"&:hover": {
+							borderColor: "@color.gray-600",
+						},
+						"&:dark": {
+							background: "@color.gray-900",
+							borderColor: "@color.gray-700",
+							color: "@color.white",
+						},
+						"&:dark:hover": {
+							borderColor: "@color.gray-600",
+						},
+					},
+				},
+				{
+					match: { color: "dark", variant: "soft" },
+					css: {
+						background: "@color.gray-800",
+						borderColor: "@color.gray-700",
+						color: "@color.gray-300",
+						"&:hover": {
+							borderColor: "@color.gray-600",
+						},
+						"&:dark": {
+							background: "@color.gray-800",
+							borderColor: "@color.gray-700",
+							color: "@color.gray-300",
+						},
+						"&:dark:hover": {
+							borderColor: "@color.gray-600",
+						},
+					},
+				},
+				{
+					match: { color: "dark", variant: "ghost" },
+					css: {
+						background: "transparent",
+						borderColor: "transparent",
+						color: "@color.gray-300",
+						"&:dark": {
+							background: "transparent",
+							borderColor: "transparent",
+							color: "@color.gray-300",
+						},
+					},
+				},
+
+				// Neutral color (adaptive: light in light mode, dark in dark mode)
+				{
+					match: { color: "neutral", variant: "default" },
+					css: {
+						background: "@color.white",
+						borderColor: "@color.gray-200",
+						color: "@color.text",
+						"&:hover": {
+							borderColor: "@color.gray-300",
+						},
+						"&:dark": {
+							background: "@color.gray-900",
+							borderColor: "@color.gray-700",
+							color: "@color.white",
+						},
+						"&:dark:hover": {
+							borderColor: "@color.gray-600",
+						},
+					},
+				},
+				{
+					match: { color: "neutral", variant: "soft" },
+					css: {
+						background: "@color.gray-100",
+						borderColor: "@color.gray-200",
+						color: "@color.text",
+						"&:hover": {
+							borderColor: "@color.gray-300",
+						},
+						"&:dark": {
+							background: "@color.gray-800",
+							borderColor: "@color.gray-700",
+							color: "@color.gray-300",
+						},
+						"&:dark:hover": {
+							borderColor: "@color.gray-600",
+						},
+					},
+				},
+				{
+					match: { color: "neutral", variant: "ghost" },
+					css: {
+						background: "transparent",
+						borderColor: "transparent",
+						color: "@color.text",
+						"&:dark": {
+							background: "transparent",
+							borderColor: "transparent",
+							color: "@color.white",
+						},
+					},
+				},
+
+				// Invalid state — overrides border and focus ring to error color.
+				{
+					match: { invalid: "true" },
+					css: {
+						borderColor: "@color.error",
+						"&:hover": {
+							borderColor: "@color.error",
+						},
+						"&:focus-within": {
+							outlineColor: "@color.error",
+						},
+						"&:dark": {
+							borderColor: "@color.error",
+						},
+						"&:dark:hover": {
+							borderColor: "@color.error",
+						},
+					},
+				},
+
+				// Read-only state — subtle background shift, default cursor.
+				{
+					match: { readonly: "true" },
+					css: {
+						background: "@color.gray-50",
+						cursor: "default",
+						"&:dark": {
+							background: "@color.gray-800",
+						},
+					},
+				},
+
+				// Disabled state — dim, not-allowed cursor, block interaction.
+				{
+					match: { disabled: "true" },
+					css: {
+						opacity: "0.5",
+						cursor: "not-allowed",
+						pointerEvents: "none",
+					},
+				},
 				...(extras.compoundVariants ?? []),
 			],
+			/** Shared default selections (everything except textarea's `resize`). */
 			defaultVariants: {
-				...fieldRecipeDefaultVariants,
+				color: "neutral",
+				variant: "default",
+				size: "md",
+				invalid: "false",
+				disabled: "false",
+				readonly: "false",
 				...extras.defaultVariants,
 			},
 		},
@@ -464,4 +452,35 @@ export function createFieldGroupRecipe(name: string) {
 			});
 		},
 	);
+}
+
+/**
+ * Registers the shared reset for the transparent native control nested inside a
+ * field wrapper (`.input-field`, `.textarea-field`). The control is transparent,
+ * borderless, and has no padding of its own &mdash; it inherits typography and
+ * color from the wrapper, which owns the visual field.
+ *
+ * The `query` argument lets each field recipe target its own nested element, so
+ * the same reset backs both the input and textarea recipes.
+ */
+export function useFieldSelector(s: Styleframe, query: string) {
+	const { selector } = s;
+
+	selector(query, {
+		flexGrow: "1",
+		minWidth: "0",
+		width: "100%",
+		background: "transparent",
+		border: "none",
+		outline: "none",
+		padding: "0",
+		color: "inherit",
+		fontFamily: "inherit",
+		fontSize: "inherit",
+		fontWeight: "inherit",
+		lineHeight: "inherit",
+		"&::placeholder": {
+			color: "@color.text-weakest",
+		},
+	});
 }
